+++ app/src/renderer/pages/app.tsx
@@ -4,7 +4,7 @@ import { useAtomValue } from 'jotai'
 import { CommandIcon, CornerDownLeftIcon as KbdEnterIcon } from 'lucide-react'
 import { sort } from 'radash'
 import { useMemo, useRef, useState } from 'react'
-import { ClientEvent, IpcRequestEventName } from '../../constants/ipc'
+import { ClientEvent, IpcRequestEventName } from '../../shared/ipc'
 import { commandsAtom } from '../atoms/data'
 import { ActionBar } from '../components/action-bar'
 import { CommandList } from '../components/command'
