+++ app/tsconfig.web.json
@@ -5,7 +5,7 @@
     "./src/renderer/**/*",
     "./src/renderer/**/*.tsx",
     "./src/preload/*.d.ts",
-    "./src/constants/**/*"
+    "./src/shared/**/*"
   ],
   "compilerOptions": {
     "composite": true,
